Port a bunch of matches from the edited branch

+++ include/core2/1EB5E70.h
@@ -5,7 +5,7 @@
 #include "core2/1EAF950.h"
 
 void* func_800DE448(f32* position, f32* arg1, f32 scale, f32* arg3, s32 arg4);
-void func_800DF410(s16 a0);
+void func_800DF410(s32 a0);
 void func_800DF41C(s32);
 void func_800DF5D8(u32, u32, u32, s32);
 void func_800DF744(s32, s32);
